Added Actor Controller
New controller added for managing requests for Actors.
